+++ include/CMakeLists.txt
@@ -1,7 +1,6 @@
 set(Includes
         # Headers
         "CelestialBody.h"
-        "MilkyWay.h"
         "SolarSystem.h"
         "Window.h"
         #".h"
