--- include/MilkyWay.h
@@ -1,10 +1,0 @@
-#ifndef MILKYWAY_H
-#define MILKYWAY_H
-
-#include <vector>
-
-#include "CelestialBody.h"
-
-std::vector<CelestialBody> getMilkyWay();
-
-#endif // MILKTYWAY_H
