+++ src/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(Sources
         # Sources
-        "MilkyWay.cpp"
+        "CelestialBody.cpp"
+        "main.cpp"
         "SolarSystem.cpp"
         "Window.cpp"
         #".cpp"
