+++ src/main.cpp
@@ -1,0 +1,16 @@
+#include <vector>
+
+#include "CelestialBody.h"
+#include "SolarSystem.h"
+#include "Window.h"
+
+int main()
+{
+  std::vector<CelestialBody> SolarSystem = getSolarSystem();
+
+  Window window;
+  while(window.Open())
+  {
+    window.Update();
+  }
+}
